requirements.txt
faiss-cpu==1.13.0
sentence-transformers==5.1.2
rank-bm25==0.2.2
jieba==0.42.1
ragas==0.1.22
datasets==4.5.0
openai==1.109.1
python-dotenv==1.2.1
torch>=2.0.0
transformers>=4.40.0
nest_asyncio
FlagEmbedding
